refactor plugins to separate files
